Code used in analysis of energy distribution of gamma rays created in a decay of ortho-positronium
The data is stored in a .root file which contains three TTree variables that each contain data for a single detector (analysis.chX), and one TTree variable that contains the other three variables (data). The calibration data is inspected at first to determine the ranges for gaussian fits, and is then used to determine the calibration graphs for each detector. Finally, graphs of all the data are drawn using the correct energy scale.