Code used to analyze and create histograms of total energies from three detectors
The data for analysis consists of three .root files. Each file has 4 TTree variables declared in them, three of them are data from a single detector each (analysis.chX), and the last one contains the other ones (data). The detectors use an arbitrary energy scale, so two of those files are calibration data. The calibration data is visually analyzed at first to determine the ranges for gaussian fits. for the known energies. The data is then passed to this program, which creates a calibration graph and finally displays histograms of all the data.